requirements.txt
pygame-ce
box2d
